+++ CHANGELOG.md
@@ -13,3 +13,7 @@ Changes since Surag started cleanup.
 9 March
 - removed random genomewide negatives
 - removed logic for background only training (releted to bias model training for ChromBPNet)
+
+14 March
+- removed getChromPositions from seqUtils
+- started adding unit tests beginning with sequtils